script/requirements.txt
asarPy==1.0.1
psycopg2-binary==2.9.10
setuptools==80.9.0
wheel==0.45.1
